Start reorganizing applogic modules
