+++ applogic/Cargo.toml
@@ -28,7 +28,7 @@ phnxtypes = { path = "../types" }
 anyhow = { version = "1.0", features = ["backtrace"] }
 serde = { version = "1", features = ["derive"] }
 tokio = { version = "1.29.0", features = ["rt", "macros"] }
-flutter_rust_bridge = { version = "=2.0.0-dev.36", features = ["chrono"] }
+flutter_rust_bridge = { version = "=2.0.0-dev.37", features = ["chrono"] }
 notify-rust = "4"
 lazy_static = "1.4"
 chrono = { version = "0.4", features = ["serde"] }
