+++ coreclient/src/clients/create_user.rs
@@ -493,14 +493,14 @@ impl PersistedUserState {
         self,
         connection: Connection,
         api_clients: ApiClients,
-    ) -> SelfUser {
+    ) -> CoreUser {
         let QsRegisteredUserState {
             key_store,
             server_url: _,
             qs_user_id,
             qs_client_id,
         } = self.state;
-        SelfUser {
+        CoreUser {
             sqlite_connection: connection,
             key_store,
             _qs_user_id: qs_user_id,
