+++ prototype/flutter_rust_bridge.yaml
@@ -7,6 +7,6 @@
 # dart_output: lib/src/rust
 
 rust_root: ../applogic
-rust_input: ../applogic/src/dart_api.rs
+rust_input: ../applogic/src/api
 dart_output: lib/core
 add_mod_to_lib: false
