+++ prototype/lib/conversation_pane/conversation_content/conversation_content.dart
@@ -6,7 +6,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:prototype/conversation_pane/conversation_content/display_message_tile.dart';
-import 'package:prototype/core/types.dart';
+import 'package:prototype/core/api/types.dart';
 import 'package:prototype/core_client.dart';
 import 'text_message_tile.dart';
 import 'package:flutter/scheduler.dart';
