+++ prototype/lib/conversation_pane/conversation_content/display_message_tile.dart
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 import 'package:flutter/material.dart';
-import 'package:prototype/core/types.dart';
+import 'package:prototype/core/api/types.dart';
 import 'package:prototype/styles.dart';
 import 'tile_timestamp.dart';
