+++ prototype/lib/conversation_pane/conversation_content/text_message_tile.dart
@@ -4,7 +4,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:prototype/conversation_pane/message_renderer.dart';
-import 'package:prototype/core/types.dart';
+import 'package:prototype/core/api/types.dart';
 import 'package:prototype/core_client.dart';
 import 'package:prototype/elements.dart';
 import 'package:prototype/styles.dart';
