+++ prototype/lib/conversation_pane/conversation_details/add_members.dart
@@ -7,7 +7,7 @@ import 'dart:async';
 import 'dart:collection';
 
 import 'package:flutter/material.dart';
-import 'package:prototype/core/types.dart';
+import 'package:prototype/core/api/types.dart';
 import 'package:prototype/core_client.dart';
 import 'package:prototype/elements.dart';
 import 'package:prototype/styles.dart';
