+++ prototype/lib/conversation_pane/conversation_details/connection_details.dart
@@ -4,7 +4,7 @@
 
 // New widget that shows conversation details
 import 'package:flutter/material.dart';
-import 'package:prototype/core/types.dart';
+import 'package:prototype/core/api/types.dart';
 import 'package:prototype/core_client.dart';
 import 'package:prototype/elements.dart';
 import 'package:prototype/styles.dart';
