+++ prototype/lib/conversation_pane/conversation_details/conversation_details.dart
@@ -8,7 +8,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:prototype/conversation_pane/conversation_details/connection_details.dart';
 import 'package:prototype/conversation_pane/conversation_details/group_details.dart';
-import 'package:prototype/core/types.dart';
+import 'package:prototype/core/api/types.dart';
 import 'package:prototype/core_client.dart';
 import 'package:prototype/elements.dart';
