+++ prototype/lib/conversation_pane/conversation_details/group_details.dart
@@ -9,7 +9,7 @@ import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:prototype/conversation_pane/conversation_details/add_members.dart';
 import 'package:prototype/conversation_pane/conversation_details/member_details.dart';
-import 'package:prototype/core/types.dart';
+import 'package:prototype/core/api/types.dart';
 import 'package:prototype/core_client.dart';
 import 'package:prototype/elements.dart';
 import 'package:prototype/styles.dart';
