+++ prototype/lib/conversation_pane/conversation_details/member_details.dart
@@ -6,7 +6,7 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:prototype/core/types.dart';
+import 'package:prototype/core/api/types.dart';
 import 'package:prototype/core_client.dart';
 import 'package:prototype/elements.dart';
 import 'package:prototype/styles.dart';
