+++ prototype/lib/conversation_pane/conversation_pane.dart
@@ -7,7 +7,7 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:prototype/conversation_pane/conversation_details/conversation_details.dart';
-import 'package:prototype/core/types.dart';
+import 'package:prototype/core/api/types.dart';
 import 'package:prototype/core_client.dart';
 import 'package:prototype/main.dart';
 import 'package:prototype/messenger_view.dart';
