+++ prototype/pubspec.yaml
@@ -28,7 +28,7 @@ dependencies:
   flutter:
     sdk: flutter
   convert: ^3.0.0
-  flutter_rust_bridge: 2.0.0-dev.36
+  flutter_rust_bridge: 2.0.0-dev.37
   freezed_annotation: ^2.2.0
   #applogic:
   #  path: ../applogic/dart-bridge
